# src/web/openapi/aethera.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Aethera API",
    "description": "REST API of the Aethera control plane. Conventions: docs/architecture/0003-api-conventions.md.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List jobs",
        "operationId": "listJobs",
        "parameters": [
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "Cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "createdAfter",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "createdBefore",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "startedAfter",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "startedBefore",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "finishedAfter",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          },
          {
            "name": "finishedBefore",
            "in": "query",
            "schema": {
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfJobDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated (auth.unauthenticated).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Role or token scope insufficient (auth.forbidden, auth.insufficient_scope).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error (internal.error).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ],
        "x-required-scope": "read"
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get a job",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated (auth.unauthenticated).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Role or token scope insufficient (auth.forbidden, auth.insufficient_scope).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error (internal.error).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ],
        "x-required-scope": "read"
      }
    },
    "/api/v1/jobs/{id}/logs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Read a job's log, paged by sequence, or download it as text",
        "description": "JSON pages by default (`fromSequence`, `limit`). With `download=true` the whole log from `fromSequence` is streamed as `text/plain`.",
        "operationId": "getJobLogs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fromSequence",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "download",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated (auth.unauthenticated).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Role or token scope insufficient (auth.forbidden, auth.insufficient_scope).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error (internal.error).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ],
        "x-required-scope": "read"
      }
    },
    "/api/v1/jobs/{id}/cancel": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel a job",
        "description": "A queued job is cancelled immediately; a running job is asked to stop and ends cancelled within one heartbeat.",
        "operationId": "cancelJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated (auth.unauthenticated).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Role or token scope insufficient (auth.forbidden, auth.insufficient_scope).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error (internal.error).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ],
        "x-required-scope": "deploy"
      }
    },
    "/api/v1/jobs/{id}/retry": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Retry a failed or cancelled job",
        "description": "Creates a new job linked to the original (`parentJobId`); the original stays as it is.",
        "operationId": "retryJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated (auth.unauthenticated).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Role or token scope insufficient (auth.forbidden, auth.insufficient_scope).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error (internal.error).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ],
        "x-required-scope": "deploy"
      }
    }
  },
  "components": {
    "schemas": {
...
      "PageOfJobDto": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobDto"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Always urn:aethera:problem:<code>."
          },
          "title": {
            "type": "string",
            "description": "Short fixed phrase per code. Do not parse."
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string",
            "description": "Occurrence-specific text for humans. Do not parse."
          },
          "instance": {
            "type": "string",
            "description": "The request path."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, area.reason. Clients branch on this. Resource-specific codes (<resource>.not_found, <resource>.already_exists) are open-ended.",
            "x-known-codes": [
              "auth.forbidden",
              "auth.insufficient_scope",
              "auth.unauthenticated",
              "concurrency.conflict",
              "confirmation.required",
              "domain.rule_violation",
              "internal.error",
              "pagination.invalid_cursor",
              "precondition.failed",
              "rate_limited",
              "request.malformed",
              "request.method_not_allowed",
              "request.not_acceptable",
              "request.too_large",
              "request.unsupported_media_type",
              "resource.conflict",
              "route.not_found",
              "service.unavailable",
              "validation.failed",
              "validation.invalid_parameter"
            ]
          },
          "traceId": {
            "type": "string",
            "description": "The X-Request-Id of the failed request."
          }
        },
        "additionalProperties": { },
        "description": "RFC 9457 problem details."
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "description": "Browser session. Unsafe methods must also send the X-CSRF-Token header (GET /auth/csrf).",
        "name": "__Host-aethera_session",
        "in": "cookie"
      },
      "bearerAuth": {
        "type": "http",
        "description": "API token: Authorization: Bearer aeth_... Its scopes are listed per operation as x-required-scope.",
        "scheme": "bearer",
        "bearerFormat": "aeth_<token>"
      }
    }
  },
...
}